Create and Delete card in real-time
