+++ backend/app/main.py
@@ -1,4 +1,4 @@
-from fastapi import Depends, FastAPI, Response, status
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse
 from sqlalchemy.orm import configure_mappers
@@ -7,9 +7,8 @@ from starlette.exceptions import HTTPException as StarletteHTTPException
 from app.config import Settings, get_settings
 from app.db.base import Base
 from app.db.database import engine
-from app.routers import category, job, user
-from app.security.auth0 import VerifyToken
-from app.security.verify_token import token_auth_scheme
+from app.routers import category, debug, job, user
+
 
 configure_mappers()
 Base.metadata.create_all(bind=engine)
@@ -36,45 +35,4 @@ async def http_exception_handler(request, exc):
 app.include_router(user.router)
 app.include_router(job.router)
 app.include_router(category.router)
-
-
-@app.get("/api/public")
-def public():
-    """No access token required to access this route"""
-
-    result = {
-        "status": "success",
-        "msg": (
-            "Hello from a public endpoint! You don't need to be "
-            "authenticated to see this."
-        ),
-    }
-    return result
-
-
-@app.get("/api/private")
-def private(response: Response, token: str = Depends(token_auth_scheme)):
-    """A valid access token is required to access this route"""
-
-    result = VerifyToken(token.credentials).verify()
-
-    if result.get("status"):
-        response.status_code = status.HTTP_400_BAD_REQUEST
-        return result
-
-    return result
-
-
-@app.get("/api/private-scoped")
-def private_scoped(response: Response, token: str = Depends(token_auth_scheme)):
-    """A valid access token and an appropriate scope are required to access
-    this route
-    """
-
-    result = VerifyToken(token.credentials, scopes="read:messages").verify()
-
-    if result.get("status"):
-        response.status_code = status.HTTP_400_BAD_REQUEST
-        return result
-
-    return result
+app.include_router(debug.router)
