+++ backend/pyproject.toml
@@ -7,7 +7,7 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = ">=3.8.1,<4.0"
-fastapi = "^0.100.1"
+fastapi = "^0.101.0"
 pydantic-settings = "^2.0.2"
 uvicorn = "^0.23.1"
 secure = "^0.3.0"
