+++ frontend/src/atoms/boardAtom.ts
@@ -1,0 +1,4 @@
+import type { Category } from '@/types/board';
+import { atom } from 'jotai';
+
+export const columnsAtom = atom<Category[]>([]);
