+++ frontend/src/components/board/Board.tsx
@@ -1,10 +1,12 @@
+'use client';
+
 import type { Category } from '@/types/board';
 import { useDroppable } from '@dnd-kit/core';
 import { SortableContext, rectSortingStrategy } from '@dnd-kit/sortable';
 import { Typography } from '@mui/material';
 import Box from '@mui/material/Box';
 import Paper from '@mui/material/Paper';
-import type { FC } from 'react';
+import { type FC } from 'react';
 import CardForm from '../popup/CardForm';
 import ActionAreaCard from './ActionAreaCard';
