+++ frontend/src/components/popup/CardForm.tsx
@@ -1,3 +1,4 @@
+import { columnsAtom } from '@/atoms/boardAtom';
 import { useAccessToken } from '@/contexts/AccessTokenContext';
 import { jobApi } from '@/services/job';
 import type { FullJobCreate } from '@/types/board';
@@ -16,6 +17,7 @@ import { DatePicker } from '@mui/x-date-pickers-pro/';
 import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import ja from 'date-fns/locale/ja';
+import { useAtom } from 'jotai';
 import moment from 'moment';
 import type { FormEvent } from 'react';
 import { useState } from 'react';
@@ -34,6 +36,7 @@ type CardFormProps = {
 const CardForm = ({ categoryId, categoryType }: CardFormProps) => {
   const { accessToken } = useAccessToken();
   const [open, setOpen] = useState(false);
+  const [columns, setColumns] = useAtom(columnsAtom);
 
   // Job information
   const [companyName, setCompanyName] = useState('');
@@ -76,9 +79,13 @@ const CardForm = ({ categoryId, categoryType }: CardFormProps) => {
       },
       selection_flows: flowProcesses,
     };
-    console.log(cardDetail);
-    jobApi.createJob(accessToken, cardDetail);
-
+    await jobApi.createJob(accessToken, cardDetail);
+    const newColumns = await jobApi.getCategoryJobs(
+      accessToken,
+      columns[0].user_id,
+      columns[0].type
+    );
+    setColumns(newColumns);
     handleClose();
   };
 
@@ -174,10 +181,7 @@ const CardForm = ({ categoryId, categoryType }: CardFormProps) => {
                 type="number"
                 inputProps={{ min: 0, max: 31, style: { fontSize: '16px' } }}
                 onChange={(e) => {
-                  // const newValue = e.target.value;
-                  // if (newValue === '' || /^[0-9]+$/.test(newValue)) {
                   setInternshipDate(e.target.value);
-                  // }
                 }}
               />
               <PeriodSelector onPeriodChange={handlePeriodChange} />
